// prisma/generated/type-graphql/resolvers/inputs/TicketExpenseWhereInput.ts
import * as TypeGraphQL from "type-graphql";
import GraphQLJSON from "graphql-type-json";
import { Prisma } from "@prisma/client";
import { DateTimeFilter } from "../inputs/DateTimeFilter";
import { FloatFilter } from "../inputs/FloatFilter";
import { IntFilter } from "../inputs/IntFilter";
import { IntNullableFilter } from "../inputs/IntNullableFilter";
import { StringFilter } from "../inputs/StringFilter";
import { TicketRelationFilter } from "../inputs/TicketRelationFilter";
import { UserRelationFilter } from "../inputs/UserRelationFilter";

@TypeGraphQL.InputType({
  isAbstract: true
})
export class TicketExpenseWhereInput {
  @TypeGraphQL.Field(_type => [TicketExpenseWhereInput], {
    nullable: true
  })
  AND?: TicketExpenseWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [TicketExpenseWhereInput], {
    nullable: true
  })
  OR?: TicketExpenseWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [TicketExpenseWhereInput], {
    nullable: true
  })
  NOT?: TicketExpenseWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  id?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => UserRelationFilter, {
    nullable: true
  })
  createdBy?: UserRelationFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  userId?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => TicketRelationFilter, {
    nullable: true
  })
  Ticket?: TicketRelationFilter | undefined;

  @TypeGraphQL.Field(_type => IntNullableFilter, {
    nullable: true
  })
  ticketId?: IntNullableFilter | undefined;

  @TypeGraphQL.Field(_type => FloatFilter, {
    nullable: true
  })
  dollars?: FloatFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  description?: StringFilter | undefined;
}
